Removed agent.py
